Modularize workspace metadata helpers
